feat: Backend init before refactoring
