+++ server/model/User.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 
-const user = new mongoose.Schema({
+const userSchema = new mongoose.Schema({
   githubId: {
     required: true,
     type: String,
@@ -17,4 +17,4 @@ const user = new mongoose.Schema({
   },
 });
 
-module.exports = mongoose.model("User", user);
+module.exports = mongoose.model("User", userSchema);
